backend/requirements.txt
flask>=3.1.0
flask-cors>=4.0.1
sqlalchemy>=2.0.36
flask-sqlalchemy>=3.1.1
marshmallow>=3.21.3
openai>=1.0.0
python-dotenv>=1.0.1
pytest>=8.2.2
pytest-flask>=1.3.0
